af: fix neon machine requiring jazelle
This will only enable Jazelle support for NEON machines built for CPUs
that have NEON. Actually, it turns out that if a CPU has NEON
(example: the cortex series) it most likely does not have Jazelle!
Jazelle is kind of the mystery of ARM that only seems to exist on
javacards and the internet is largely lacking in knowledge of how to use
it well. Probably the only people that use it and know how to use it
work on javacards.